requirements.txt
aiohttp==3.12.12
aiosignal==1.3.2
amqp==5.3.1
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
asynctest==0.13.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
billiard==4.2.1
bleach==6.2.0
cached-property==2.0.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
crispy-bootstrap5==2025.6
cron-descriptor==1.4.5
cryptography>=41.0.0
daphne==4.2.0
decouple==0.0.7
defusedxml==0.7.1
dj-database-url==3.0.0
dj-stripe==2.9.1
Django==5.2.3
django-allauth==65.9.0
django-appconf==1.1.0
django-birthday==0.1.4
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-countries==7.6.1
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
django-environ==0.12.0
django-filter==25.1
django-flexible-subscriptions==0.15.1
django-heroku==0.3.1
django-ipware==7.0.1
django-jazzmin==3.0.1
django-jsonfield==1.4.1
django-markdownify==0.9.5
django-mobile==0.7.0
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-notify==1.1.2
django-paypal==2.1
django-polymorphic==4.1.0
django-redis==5.4.0
django-simple-history==3.8.0
django-sslserver==0.22
django-templated-mail==1.1.1
django-timezone-field==7.1
django-tracking-analyzer==1.1.1
django-user-agents==0.4.0
django-utils-six==2.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework-jsonapi==7.1.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.3.0
fastjsonschema==2.21.1
fqdn==1.5.1
frozenlist==1.7.0
geoip2==5.1.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==24.7.2
inflection==0.5.1
isoduration==20.11.0
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonfield==3.1.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
Markdown==3.8
MarkupSafe==3.0.2
maxminddb==2.7.0
mistune==3.1.3
msgpack==1.1.0
multidict==6.4.4
nest-asyncio==1.6.0
numpy==2.2.6
oauthlib==3.2.2
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pillow==11.2.1
pkgutil_resolve_name==1.3.10
platformdirs==4.3.8
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-gettext==5.0
python-ipware==3.0.0
python-json-logger==3.3.0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.15.3
secure_bite==0.1.3
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
stripe==11.6.0
swapper==1.4.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.2.1
tornado==6.5.1
traitlets==5.14.3
Twisted==25.5.0
txaio==23.1.1
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.4.0
user-agents==2.2.0
uvicorn==0.34.3
vine==5.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.9.0
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zope.interface==7.2
